pyproject.toml
[project]
name = "data_shield"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "mypy>=1.15.0",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.4",
    "ruff>=0.9.9",
    "specklepy>=2.21.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "UP", # pyupgrade
    "D", # pydocstyle
    "I", # isort
]
line-length = 120
ignore = ["F401", "F403"]

[tool.ruff.pydocstyle]
convention = "google"
